+++ app/src/test/kotlin/me/xizzhu/android/joshua/reading/detail/CrossReferenceItemTest.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright (C) 2022 Xizhi Zhu
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package me.xizzhu.android.joshua.reading.detail
+
+import me.xizzhu.android.joshua.R
+import me.xizzhu.android.joshua.core.StrongNumber
+import me.xizzhu.android.joshua.core.Verse
+import me.xizzhu.android.joshua.core.VerseIndex
+import me.xizzhu.android.joshua.tests.BaseUnitTest
+import me.xizzhu.android.joshua.tests.MockContents
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertTrue
+
+@RunWith(RobolectricTestRunner::class)
+class CrossReferenceItemTest : BaseUnitTest() {
+    @Test
+    fun testItemViewType() {
+        assertEquals(R.layout.item_cross_reference, CrossReferenceItem(VerseIndex.INVALID, Verse.Text("", ""), "").viewType)
+    }
+
+    @Test
+    fun testTextForDisplay() {
+        assertEquals(
+                "Genesis 1:1\nIn the beginning God created the heaven and the earth.",
+                CrossReferenceItem(VerseIndex(0, 0, 0), MockContents.kjvVerses[0].text, MockContents.kjvBookNames[0]).textForDisplay.toString()
+        )
+    }
+}