+++ app/src/test/kotlin/me/xizzhu/android/joshua/search/SearchViewModelTest.kt
@@ -420,4 +420,40 @@ class SearchViewModelTest : BaseUnitTest() {
         assertEquals(3, (actual[1] as BaseViewModel.ViewData.Success).data.items.size)
         assertEquals(1, (actual[1] as BaseViewModel.ViewData.Success).data.currentPosition)
     }
+
+
+    @Test
+    fun `test toSearchVersePreviewItems() with single verse`() {
+        searchViewModel.searchRequest.value = SearchViewModel.SearchRequest("God", true)
+        val actual = with(searchViewModel) { toSearchVersePreviewItems(listOf(MockContents.kjvVerses[0])) }
+        assertEquals(1, actual.size)
+        assertEquals("1:1 In the beginning God created the heaven and the earth.", actual[0].textForDisplay.toString())
+    }
+
+    @Test
+    fun `test toSearchVersePreviewItems() with multiple verses`() {
+        searchViewModel.searchRequest.value = SearchViewModel.SearchRequest("God", true)
+        val actual = with(searchViewModel) { toSearchVersePreviewItems(listOf(MockContents.kjvVerses[0], MockContents.kjvVerses[1])) }
+        assertEquals(2, actual.size)
+        assertEquals("1:1 In the beginning God created the heaven and the earth.", actual[0].textForDisplay.toString())
+        assertEquals(
+                "1:2 And the earth was without form, and void; and darkness was upon the face of the deep. And the Spirit of God moved upon the face of the waters.",
+                actual[1].textForDisplay.toString()
+        )
+    }
+
+    @Test
+    fun `test toSearchVersePreviewItems() with multiple verses but not consecutive`() {
+        searchViewModel.searchRequest.value = SearchViewModel.SearchRequest("God", true)
+        val actual = with(searchViewModel) { toSearchVersePreviewItems(listOf(MockContents.msgVerses[0], MockContents.msgVerses[1], MockContents.msgVerses[2])) }
+        assertEquals(2, actual.size)
+        assertEquals(
+                "1:1-2 First this: God created the Heavens and Earth—all you see, all you don't see. Earth was a soup of nothingness, a bottomless emptiness, an inky blackness. God's Spirit brooded like a bird above the watery abyss.",
+                actual[0].textForDisplay.toString()
+        )
+        assertEquals(
+                "1:3 God spoke: \"Light!\"\nAnd light appeared.\nGod saw that light was good\nand separated light from dark.\nGod named the light Day,\nhe named the dark Night.\nIt was evening, it was morning—\nDay One.",
+                actual[1].textForDisplay.toString()
+        )
+    }
 }