+++ docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 CHANGELOG
 ---------
 
+#### Next release
+- New features:
+  - Support cross references (#21)
+
 #### v0.20.2 (2022-02-07)
 - Bug fixes:
   - Should call AppCompatDelegate.setDefaultNightMode() on main thread